feat: rewrite simplified variant using claude 4 thinking
Don't use any markdown.
Don't use any webcomponents.
Start with a single self-contained html file (except for `secrets.js`.

Main prompt:

```
The purpose of this app is to allow people to access information if something happens to me.
I will also include my passwords and sensitive information and therefore intend to protect this data by requiring a certain number of persons to agree that it is time to access my data.  If and only if a certain number enter their passwords at the same time.  They will be able to access the content.  This should already be implemented using the shamir algorithm.
I want to send a single file to my family friends which should include all necessary data and not depend on any external links.  This is to make sure that the page works even in for instance 30 years.
The app should also allow me to easily update the private content.  This means that the app must be able to create a downloadable version of itself with updated content.
I intend to provide different information for different use cases.  This allows me to differentiate between the case, where I am dead versus the case where I am simply temporarily incapacitated.
Content in this context could mean multiple files.  I want to be able to upload scanned documents or pdfs as well.  The current code zips those files, but I think that the zip functionality is overkill and should be removed.  Currently I can enter text using a markdown editor.  This is overkill as well and should be removed.  Content should therefore be uploaded files, together with an unencrypted description and a text (which should be encrypted together with the uploaded files) associated with it.  (If it makes code easier, we could store the associated text as README.TXT next to the uploaded files)

If somebody opens the page it should by default explain how to use the page (note that it shouldn't explain what the page is for, but how to use it.  I want to use it for my final instructions, but somebody else might for instance wan't to use it for company secrets.) and list the available contents.  (If possible make it easy to translate the whole app; follow best practices).  Every content should have a short (unencrypted) description and information about how many passwords are necessary.
The user should then select one of them and the web page should then ask for the passwords (displaying a "decryption" progress bar).
After decryption the associated text together with the list of files should be shown.The browser should simply show download links for the uploaded files (even if they are other text files).  If the browser displays them directly after clicking on it, that's fine.

I also need an admin interface.  If I open the webpage with an added `#admin`. The admin area should be shown.

It should allow me also list all available contents.  Let me edit or delete a content or create a new content.  To access existing content I will obviously also need to enter a password.
Creating new content should happen in multiple steps.  I will first need to enter all information necessary for the passwords: how many passwords should be created, and how many passwords will be necessary for decryption.  In this step I will also need to enter a name for the content, so that the password creation step can attach this name to the generated passwords (which will make it easier for the user and allow the decryption step to verify that the correct password is used.)  This step will also need to give me a "master" password.  This could be either the "real" de-/encryption key (base64 encoded) or a string which allows us to extract n passwords as if users entered them (and use the sharmi algorithm to recreate the "real" key).  (See already existing code).  I have no preference.  Use what you think makes more sense or is better practice.

If I edit an existing content it would be nice to be able to see the keys I sent to users.  This would allow me to resend a key to a person, if they lost their key.  (Maybe store a mapping of the keys with names/mail adresses in the content itself?)

The current implementation is already pretty far, but unfinished.

I also intend to open source this application.  Please follow best practices and try to be consistent.
My current code tries to use webcomponents.
I now think that webcomponents are overkill.  Please start a completely new page and only use the existing code as inspiration.
I do however think that mixing code and html should usually be avoided and template html code is clearer.  Please really follow best practices here.

The application will need to be self-contained, but for now depend on the secrets.js.  Add a comment where I should inline its content later on to make the app really self contained.

Note that being self-contained but still modify itself is not that easy!  Whenever the admin area changes something it will need to be able to get the "initial" page, and only then add the contents.

Following best practices, being consistent and structuring the code nicely is really important for me!

Please help me create a simpler, complete version of this app, which implements my requirements.

```